Fixed the answer display
There was an issue where system CLS took place and didn't display the answer. That was fixed and a better terminal printout is now present.

There's also a error catcher. If something outside the (current) 4 functions was put in, the program will tell you to put in the correct input.